[cassandra] Make listen backlog configurable
Summary:
Java sets a default listen backlog of 50, which is too low for
high connection rates and causes overflows. This makes the listen
backlog for the hsha and async servers configurable using the existing
`rpc_listen_backlog` config parameter.

This can only be done in 2.1 because the thrift version in 2.0 doesn't
support setting the backlog.

Test Plan:
Installed on cassandraproxyhomefeed45-ash and enabled hsha.

    [root@cassandraproxyhomefeed45-ash <lxc> ~ ] # ss -ln | grep 9160
    LISTEN     2      511                      :::9160 :::*

Reviewers: dikang

Subscribers:

Differential Revision: https://phabricator.fb.com/D2707030

Tasks:

Blame Revision: