Wait gossip opening connection settle during startup
Summary:
To mark a node alive, gossiper sends an ECHO message to that node and
wait for response. Because building a SSL incoming connection takes long
time (100ms - 200ms per node) and it's single threaded. For a large
cluster, the startup node needs several minutes to build all the
connections.

We added a wait for the new incoming connection to be settled, but the
new connection building may take > 10 seconds (mostly the target node is
not responsing, or any network problem), which causes false positive.

This change checks the gossip opening connection by monitoring ongoing
ECHO message.